🚚Rename: hooks 폴더 이름 변경, SameTag pages 폴더로 이동
